Remove overflow style manipulation on body
If menu is triggered on a page higher than the viewport, setting overflow: hidden on the body will make the content 'jump' to the right on operating systems that make the scrollbar take up space in the browser (i.e. Windows).
Following up on the previous commit, this is no longer necessary.